# backend/tests/expected/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "additionalProperties": true,
        "description": "A bare ``{\"ok\": true}`` mutation acknowledgement.",
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "PlaylistCreate": {
        "properties": {
          "estimated_minutes": {
            "default": 30,
            "title": "Estimated Minutes",
            "type": "integer"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PlaylistItemInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planet": {
            "default": "Calypso",
            "title": "Planet",
            "type": "string"
          },
          "quest_ids": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Quest Ids",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "PlaylistCreate",
        "type": "object"
      },
...
      "Quest": {
        "additionalProperties": true,
        "description": "A quest as returned by the quest read/write endpoints.\n\nMirrors the frontend ``Quest`` type field-for-field, camelCase verbatim.\nEvery key is always present; nullable fields carry an explicit null.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "chainName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chainname"
          },
          "chainPosition": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chainposition"
          },
          "chainTotal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chaintotal"
          },
          "cooldownDurationHours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldowndurationhours"
          },
          "cooldownExpiresAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownexpiresat"
          },
          "expectedRewardMarkupPercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedrewardmarkuppercent"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "title": "Notes",
            "type": "string"
          },
          "planet": {
            "title": "Planet",
            "type": "string"
          },
          "playlistIds": {
            "items": {
              "type": "string"
            },
            "title": "Playlistids",
            "type": "array"
          },
          "reward": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reward"
          },
          "rewardDescription": {
            "title": "Rewarddescription",
            "type": "string"
          },
          "rewardIsSkill": {
            "title": "Rewardisskill",
            "type": "boolean"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "targetMobs": {
            "items": {
              "type": "string"
            },
            "title": "Targetmobs",
            "type": "array"
          },
          "waypoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoint"
          }
        },
        "required": [
          "id",
          "name",
          "targetMobs",
          "planet",
          "rewardIsSkill",
          "rewardDescription",
          "notes",
          "playlistIds"
        ],
        "title": "Quest",
        "type": "object"
      },
...
      "QuestCancel": {
        "properties": {
          "undo_reward": {
            "default": false,
            "title": "Undo Reward",
            "type": "boolean"
          }
        },
        "title": "QuestCancel",
        "type": "object"
      },
      "QuestCreate": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "chain_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Name"
          },
          "chain_position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Position"
          },
          "chain_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Total"
          },
          "cooldown_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldown Hours"
          },
          "expected_reward_markup_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Reward Markup Percent"
          },
          "mobs": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Mobs",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "planet": {
            "default": "Calypso",
            "title": "Planet",
            "type": "string"
          },
          "reward_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reward Description"
          },
          "reward_is_skill": {
            "default": false,
            "title": "Reward Is Skill",
            "type": "boolean"
          },
          "reward_ped": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reward Ped"
          },
          "waypoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoint"
          }
        },
        "required": [
          "name"
        ],
        "title": "QuestCreate",
        "type": "object"
      },
...
      "QuestPlaylist": {
        "additionalProperties": true,
        "description": "A playlist as returned by the playlist read/write endpoints.",
        "properties": {
          "estimatedMinutes": {
            "title": "Estimatedminutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "immediateQuestIds": {
            "items": {
              "type": "string"
            },
            "title": "Immediatequestids",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlaylistItem"
            },
            "title": "Items",
            "type": "array"
          },
          "longHorizonQuestIds": {
            "items": {
              "type": "string"
            },
            "title": "Longhorizonquestids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planet": {
            "title": "Planet",
            "type": "string"
          },
          "questIds": {
            "items": {
              "type": "string"
            },
            "title": "Questids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "planet",
          "estimatedMinutes",
          "questIds",
          "immediateQuestIds",
          "longHorizonQuestIds",
          "items"
        ],
        "title": "QuestPlaylist",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EntropiaOrme API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/quests": {
      "get": {
        "description": "List all active quests.",
        "operationId": "list_quests_api_quests_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Quest"
                  },
                  "title": "Response List Quests Api Quests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Quests",
        "tags": [
          "quests"
        ]
      },
      "post": {
        "description": "Create a new quest.",
        "operationId": "create_quest_api_quests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Quest",
        "tags": [
          "quests"
        ]
      }
    },
...
    "/api/quests/playlists": {
      "get": {
        "description": "List all active playlists.",
        "operationId": "list_playlists_api_quests_playlists_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestPlaylist"
                  },
                  "title": "Response List Playlists Api Quests Playlists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Playlists",
        "tags": [
          "quests"
        ]
      },
      "post": {
        "description": "Create a new playlist.",
        "operationId": "create_playlist_api_quests_playlists_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestPlaylist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Playlist",
        "tags": [
          "quests"
        ]
      }
    },
...
    "/api/quests/playlists/{playlist_id}": {
      "delete": {
        "description": "Soft-delete a playlist.",
        "operationId": "delete_playlist_api_quests_playlists__playlist_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Playlist",
        "tags": [
          "quests"
        ]
      },
      "put": {
        "description": "Update a playlist.",
        "operationId": "update_playlist_api_quests_playlists__playlist_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "playlist_id",
            "required": true,
            "schema": {
              "title": "Playlist Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaylistUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestPlaylist"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Playlist",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/{quest_id}": {
      "delete": {
        "description": "Soft-delete a quest.",
        "operationId": "delete_quest_api_quests__quest_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Quest",
        "tags": [
          "quests"
        ]
      },
      "get": {
        "description": "Get a single quest.",
        "operationId": "get_quest_api_quests__quest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Quest",
        "tags": [
          "quests"
        ]
      },
      "put": {
        "description": "Update a quest.",
        "operationId": "update_quest_api_quests__quest_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Quest",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/{quest_id}/cancel": {
      "post": {
        "description": "Cancel a started quest or undo an active cooldown reset.",
        "operationId": "cancel_quest_api_quests__quest_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/QuestCancel"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Quest",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/{quest_id}/complete": {
      "post": {
        "description": "Complete a quest — resets cooldown, increments counter, auto-creates ledger entry.",
        "operationId": "complete_quest_api_quests__quest_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Quest",
        "tags": [
          "quests"
        ]
      }
    },
    "/api/quests/{quest_id}/start": {
      "post": {
        "description": "Mark a quest as in-progress.",
        "operationId": "start_quest_api_quests__quest_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "quest_id",
            "required": true,
            "schema": {
              "title": "Quest Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Quest",
        "tags": [
          "quests"
        ]
      }
    },
...
  }
}